aws-lambda/requirements.txt
appdirs==1.4.3
args==0.1.0
astroid==2.4.1
attrs==19.3.0
awscrt==0.5.13
awsiotsdk==1.0.6
beautifulsoup4==4.7.1
black==19.10b0
botocore==1.16.2
bs4==0.0.1
certifi==2019.6.16
cfscrape==2.0.7
chalice==1.14.0
chardet==3.0.4
click==7.1.2
clint==0.5.1
colorama==0.4.1
docutils==0.15.2
enum-compat==0.0.3
enum34==1.1.6
future==0.17.1
idna==2.8
img2pdf==0.3.3
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
paho-mqtt==1.5.0
pathspec==0.8.0
Pillow==6.2.2
pylint==2.5.2
python-dateutil==2.8.1
regex==2020.4.4
requests==2.22.0
six==1.14.0
soupsieve==1.9.1
toml==0.10.0
tqdm==4.32.2
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
